refactor: prefer `nullptr` over `NULL` (yeah...)
I'm torn because `NULL` honestly looks way nicer than `nullptr`, partly because the `ptr` prefix is both extra length (presumably just so it wouldn't collide with any existing `null` macro? otherwise I have no idea why), and partly because the syntax highlighting chooses a different colour for `NULL` as a macro than `nullptr` as a keyword. These are really dumb reasons but hey. That's why I used `NULL` for a bit. Going back to `nullptr` though for the sake of compliant C++... though it isn't really like it matters in this codebase lol